fileUploader: add React component
This component take the form of a button "+". It allows a file to be
uploaded to the host. It takes as input, the extension of the file
allowed to be uploaded and a callback function to send information to
the parent component: name of the file uploaded, progress of the upload.

Signed-off-by: Tanguy Raufflet <tanguy.raufflet@savoirfairelinux.com>