VMActions: fix the list nodes for the migration
Nodes with "-" in the name were not displayed in the list of available
nodes to migrate to.

This commit modifies the regex to obtain the complete list of online
nodes. In addition, the includes() method used to remove the current
node from the list of nodes to migrate to has been replaced by a more
robust solution.

Signed-off-by: Tanguy Raufflet <tanguy.raufflet@savoirfairelinux.com>