VMActions: Improve VM table updating after migration
This commit removes the correlation between the table of VMs displayed,
and the pooling performed to update the status of a VM after its
migration. The previous version updated the view to obtain VM status
every 3 seconds, causing the VM table to blink.

Pooling is now performed every second, but the view is only updated if
the VM is in the "Started" state or if 15 seconds have elapsed.

Signed-off-by: Tanguy Raufflet <tanguy.raufflet@savoirfairelinux.com>