VMConsole: add React component
This new component displays the terminal of a VM running on the cluster.
It is based on the component Terminal provided by Cockpit which requires
the xterm dependencies added to the package.json.

It takes as input the name of the VM taken in page URL and the user
(libvirt user) used to connect to the VM via ssh. These two parameters
can be modified afterward in the webpage.

Signed-off-by: Tanguy Raufflet <tanguy.raufflet@savoirfairelinux.com>