VMActions: add updateSelectedVM function
After a first VM migration, the list of available nodes to perform a
second migration was not updated. Although the VM table displayed was
updated, the object contained in the props selectedVM was not.

The update of the selectedVM variable is now done after each VM list
update. Because refreshVMList() is an asynchronous function and need to
be executed before updateSelectedVM(), the async/await mechanism has
been added when needed.

Signed-off-by: Tanguy Raufflet <tanguy.raufflet@savoirfairelinux.com>