VMCreator: add option to configure the disk bus type in the VM
The Cluster VM management cockpit plugin uses vm-manager, which creates
in the VM XML configuration the entry for the VM disk.
Up until now, the default bus type for the VM disk was always "virtio".

Commit [1] of vm-manager added an option to choose the disk bus type.
Add a corresponding option in the VM creator to match this new option.

[1]: https://github.com/seapath/vm_manager/commit/7755c90e2f2fcb3eb242bd58801abf7a6d139753

Signed-off-by: Elinor Montmasson <elinor.montmasson@savoirfairelinux.com>